Reorganize framework access-control files into access/ package
Move actor.py, authz.py, capabilities.py (and their tests) from the
framework root into src/framework/access/{actor,authz,capabilities}/,
each as its own sub-package with colocated test file. Update all import
sites across domain and framework layers.
