+++ src/domain/specs/test_user.py
@@ -13,8 +13,8 @@ from src.domain.logic.users.schema import (
     UserActivationUpdate,
 )
 from src.domain.specs.user import USER_ENTITY
+from src.framework.access.authz.authz import is_self_or_admin
 from src.framework.audit.core import AuditAction
-from src.framework.authz import is_self_or_admin
 from src.framework.dispatch.entity_spec import ADMIN_FOR_WRITE, RelatedListSubresource
 
 # --- Auth deps (security-visible) ----------------------------------------
