+++ src/framework/access/capabilities/capabilities.py
