+++ src/framework/dispatch/mounts/form.py
@@ -13,7 +13,7 @@ from uuid import UUID
 from fastapi import Request
 from pydantic import BaseModel  # noqa: F401 — re-exported for test compat
 
-from src.framework.actor import Actor
+from src.framework.access.actor.actor import Actor
 from src.framework.dispatch.mounts._common import (
     assert_kind_lock,
     call_handler_with,
