+++ src/framework/dispatch/mounts/update.py
@@ -12,7 +12,7 @@ from uuid import UUID
 from fastapi import Request
 from pydantic import BaseModel
 
-from src.framework.actor import Actor
+from src.framework.access.actor.actor import Actor
 from src.framework.audit.core import mutate
 from src.framework.audit.repository import AuditRepository
 from src.framework.dispatch.mounts._common import (
